feat: move HasPermissionRule to course_roles app
