+++ lms/djangoapps/courseware/rules.py
@@ -15,7 +15,6 @@ from opaque_keys.edx.django.models import CourseKeyField
 from opaque_keys.edx.keys import CourseKey, UsageKey
 from xblock.core import XBlock
 
-from openedx.core.djangoapps.course_roles.helpers import course_or_organization_permission_check
 from openedx.core.djangoapps.content.course_overviews.models import CourseOverview
 from openedx.core.djangoapps.enrollments.api import is_enrollment_valid_for_proctoring
 from common.djangoapps.student.models import CourseAccessRole
@@ -178,25 +177,4 @@ class HasRolesRule(Rule):  # lint-amnesty, pylint: disable=abstract-method, miss
                 return True
             if OrgRole(role, course_key.org).has_user(user):
                 return True
-        return False
-
-
-class HasPermissionRule(Rule):  # lint-amnesty, pylint: disable=abstract-method, missing-class-docstring
-    def __init__(self, permission):
-        self.permission = permission
-
-    def check(self, user, instance=None):
-        if not user.is_authenticated:
-            return False
-        if isinstance(instance, CourseKey):
-            course_key = instance
-        elif isinstance(instance, (CourseBlock, CourseOverview)):
-            course_key = instance.id
-        elif isinstance(instance, (ErrorBlock, XBlock)):
-            course_key = instance.scope_ids.usage_id.course_key
-        else:
-            course_key = CourseKey.from_string(str(instance))
-
-        if course_or_organization_permission_check(user, self.permission, course_key, course_key.org):
-            return True
         return False
